Migrate to typescript - remove webpack
