--- .eslintignore
@@ -1,9 +1,0 @@
-node_modules/
-.npm
-.eslintcache
-.env
-.cache
-.next
-dist
-dist-pkg
-.DS_Store
