+++ Containerfile
@@ -1,15 +1,17 @@
-FROM node:latest
-
-ENV PUPPETEER_SKIP_DOWNLOAD=true
-
-WORKDIR /src
-
+# Build stage
+FROM node:20-alpine AS build
+WORKDIR /usr/src/app
 COPY package*.json ./
-
 RUN npm install
-
 COPY . .
+RUN npm run build
 
+# Server stage
+FROM node:20-alpine
+WORKDIR /usr/src/app
+COPY --from=build /usr/src/app/dist ./dist
+COPY --from=build /usr/src/app/package.json ./
+COPY --from=build /usr/src/app/node_modules ./node_modules
+COPY --from=build /usr/src/app/public ./public
 EXPOSE 3000
-
-CMD npm start
+CMD ["node", "./dist/server.js"]
