+++ example-daphine.yml
@@ -63,8 +63,6 @@ spec:
           ports:
             - containerPort: 3000
           env:
-            - name: PUPPETEER_SKIP_DOWNLOAD
-              value: "true"
             # Frontend url
             - name: CORS_ORIGIN
               value: "https://example.com"
