+++ package.json
@@ -1,40 +1,49 @@
 {
-  "name": "daphine_piaget",
-  "version": "0.1.2",
-  "description": "The api for freedom.",
+  "name": "daphine",
+  "version": "0.2.0",
+  "description": "The API for Freedom",
+  "author": "Jordon Leach",
+  "license": "GPL-3.0-only",
+  "main": "server.js",
   "private": true,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/jordojordo/daphine.git"
+  },
   "scripts": {
-    "start": "node ./bin/www",
-    "start:prod": "node ./dist/index.bundle.js",
-    "build": "webpack -- --mode=production"
+    "build": "tsc && copyfiles -u 1 src/views/**/* dist",
+    "start": "node ./dist/server.js",
+    "start:dev": "tsc-watch --onSuccess \"npm run serve\"",
+    "serve": "npm run copy-views && npm run start",
+    "copy-views": "copyfiles -u 1 src/views/**/* dist",
+    "lint": "npx eslint .",
+    "format": "npx eslint . --fix"
   },
   "dependencies": {
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "debug": "^4.3.4",
-    "express": "^4.18.1",
-    "express-fileupload": "^1.4.0",
-    "file-type": "^19.0.0",
-    "http-errors": "^2.0.0",
-    "morgan": "^1.10.0",
-    "path": "^0.12.7",
-    "pug": "^3.0.2",
-    "puppeteer": "^21.0.0"
+    "module-alias": "^2.2.3"
   },
   "devDependencies": {
-    "@types/node": "^20.2.5",
-    "@types/webpack": "^5.28.1",
-    "@types/webpack-dev-server": "^4.7.2",
-    "@types/webpack-node-externals": "^3.0.0",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "babel-loader": "^9.1.2",
-    "babel-preset-env": "^1.7.0",
-    "eslint": "^8.42.0",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.3",
-    "webpack": "^5.85.1",
-    "webpack-cli": "^5.1.3",
-    "webpack-node-externals": "^3.0.0"
+    "@types/cookie-parser": "^1.4.7",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/morgan": "^1.9.9",
+    "@types/node": "^20.12.5",
+    "cookie-parser": "^1.4.6",
+    "copyfiles": "^2.4.1",
+    "cors": "^2.8.5",
+    "ejs": "^3.1.9",
+    "eslint": "^8.57.0",
+    "express": "^4.19.2",
+    "globals": "^15.0.0",
+    "http-errors": "^2.0.0",
+    "morgan": "^1.10.0",
+    "puppeteer": "^22.6.3",
+    "tsc-watch": "^6.2.0",
+    "typescript": "^5.4.4"
+  },
+  "_moduleAliases": {
+    "@src": "dist",
+    "@routes": "dist/routes",
+    "@api": "dist/api"
   }
 }
