--- public/stylesheets/style.css
@@ -1,18 +1,0 @@
-body {
-  padding: 50px;
-  font: 14px "Lucida Grande", Helvetica, Arial, sans-serif;
-}
-
-a {
-  color: #00B7FF;
-}
-
-#error {
-  height: auto;
-  min-height: 100vh;
-}
-
-#error img {
-  max-width: 500px;
-  margin: 20px 0;
-}
