+++ src/routes/upload.ts
@@ -1,0 +1,9 @@
+// import express, { Request, Response } from 'express';
+// import { createUpload } from '@api/upload';
+
+// const router = express.Router();
+
+// export default router.post('/', (req: Request, res: Response) => {
+// 	createUpload(req, res);
+// });
+
