+++ src/views/error.ejs
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title><%= status %> Error</title>
+</head>
+<body>
+    <img src="<%= imageUrl %>" alt="<%= status %> Error">
+    <p><%= message %></p>
+</body>
+</html>
