--- src/views/error.pug
@@ -1,6 +1,0 @@
-extends layout
-
-block content
-  div#error
-    img(src=`https://http.cat/${ error.status }`)
-    pre #{error.stack}
