+++ tsconfig.json
@@ -1,14 +1,17 @@
 {
   "compilerOptions": {
+    "target": "ESNext",
+    "module": "commonjs",
+    "baseUrl": "./",
     "outDir": "./dist",
-    "allowJs": true,
-    "target": "ES2018",
-    "esModuleInterop": true
-  },
-  "include": [
-    "./src/**/*"
-  ],
-  "exclude": [
-    "node_modules"
-  ]
+    "paths": {
+      "@src/*": ["./src/*"],
+      "@routes/*": ["./src/routes/*"],
+      "@api/*": ["./src/api/*"]
+    }, 
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true
+  }
 }
